requirements.txt
# 核心业务依赖（适配Python 3.11，无版本冲突）
# 核心业务依赖 (放宽限制以提升服务器环境兼容性)
mem0ai>=1.0.4
neo4j>=5.0.0
langchain-core
langchain-community
langchain-experimental
langchain_ollama
ollama==0.6.1
qdrant-client==1.17.0

# 辅助依赖（代码中用到的基础库）
pydantic==2.12.5
httpx==0.27.2
protobuf==5.29.6
sqlalchemy==2.0.47
numpy<2.0.0
shutilwhich==1.1.0
tqdm==4.67.3
python-dotenv>=1.0.0
